全配置BOM一级清单前端完成特性显示

+++ src/main/java/com/connor/hozon/bom/bomSystem/dao/cfg/HzFullCfgMainDao.java
@@ -8,7 +8,7 @@ public interface HzFullCfgMainDao {
 
     int deleteByPrimaryKey(BigDecimal id);
 
-    int insert(HzFullCfgMain record);
+    BigDecimal insert(HzFullCfgMain record);
 
     int insertSelective(HzFullCfgMain record);
